rename selector package
