+++ internal/thelma/cli/selector/filter_builder.go
